pyproject.toml
[project]
name = "memory-poc"
version = "0.1.0"
description = "Temporal + Semantic + Entity Memory System for AI agents using PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "nltk>=3.8.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "rich>=13.0.0",
    "spacy>=3.7.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "langchain-text-splitters>=0.3.0",
    "flask>=3.1.2",
    "fastapi[standard]>=0.120.3",
    "uvicorn>=0.38.0",
]
